Restructure contd.
